refactor: handler.ts 폴더 내로 이동
