+++ src/mocks/server.ts
@@ -1,6 +1,6 @@
 import { setupServer } from 'msw/node';
 
-import handlers from './handlers/products';
+import handlers from './handlers/index';
 
 const server = setupServer(...handlers);
